backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pyodbc==5.3.0
pydantic==2.10.3
python-dotenv==1.0.1
azure-storage-blob==12.24.0
openai==1.58.1
google-genai==1.0.0
anthropic==0.40.0
websockets==14.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
